requirements.txt
# Housekeeping
numpy==1.26.2
pandas==2.1.3
pyarrow==14.0.1
gunicorn==21.2.0
ipykernel==6.27.1
pydantic==2.5.3
# Visualizations
orjson==3.9.10
plotly==5.18.0
dash==2.14.1
dash-bootstrap-components==1.5.0
dash-mantine-components==0.12.1
dash-iconify==0.1.2
dash-breakpoints==0.1.0
dash-gif-component==1.1.0
# Google cloud platform
google-cloud-bigquery==3.14.1
google-cloud-aiplatform==1.38.1
pandas_gbq==0.20.0